+++ src/engine/convert.h
@@ -16,9 +16,9 @@ namespace YAML {
                 return false;
             }
 
-            rhs.x_ = node[0].as<float>();
-            rhs.y_ = node[1].as<float>();
-            rhs.z_ = node[2].as<float>();
+            rhs.x_ = node[0].as<double>();
+            rhs.y_ = node[1].as<double>();
+            rhs.z_ = node[2].as<double>();
             return true;
         }
     };
