+++ src/objects/solids/Sphere.h
@@ -9,13 +9,13 @@
 
 class Sphere : public Solid {
 public:
-    Sphere(const Point& pos, TexMat& tex, float r) : Solid(pos, tex), r_(r) {}
-    float intersects(Line const& line) const override;
+    Sphere(const Point& pos, TexMat& tex, double r) : Solid(pos, tex), r_(r) {}
+    double intersects(Line const& line) const override;
     Line get_normal(Point const& p) const override;
     TexPixel get_tex(Point const& p) const override;
 
 protected:
-    float r_;
+    double r_;
 };
 
 #endif //RAYTRACER_SPHERE_H
